modify string_nconcat.c file

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -15,7 +15,7 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 char *result;
-int len1, len2;
+size_t len1, len2;
 if (s1 == NULL)
 {
 s1 = "";
